# src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Payment Initiation API - PaymentOrder
  description: |
    API REST para gestión de órdenes de pago alineada con BIAN Service Domain Payment Initiation 12.0.
    
    **Service Domain**: Payment Initiation  
    **Behavior Qualifier**: PaymentOrder  
    **Control Record**: PaymentOrderProcedure
    
    Esta API implementa nomenclatura BIAN con entidades anidadas según el BOM Diagram oficial.
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: dev@bank.com
  license:
    name: Proprietary

servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
  - url: https://api.bank.com
    description: Servidor de producción

tags:
  - name: Payment Orders
    description: Operaciones de órdenes de pago según BIAN Payment Initiation

paths:
  /payment-initiation/payment-orders:
    post:
      tags:
        - Payment Orders
      summary: Iniciar orden de pago
      description: |
        Crea una nueva orden de pago (Initiate PaymentOrder).
        Mapea la operación SOAP SubmitPaymentOrder a REST con estructura BIAN.
      operationId: initiatePaymentOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiatePaymentOrderRequest'
            examples:
              example1:
                summary: Orden de pago estándar
                value:
                  paymentOrderProcedure:
                    paymentTransactionInitiatorReference: "EXT-123"
                    payer:
                      payerReference: "Juan Pérez"
                      payerBankReference: "BANK001"
                      payerProductInstanceReference: "EC12DEBTOR"
                    payee:
                      payeeReference: "María López"
                      payeeBankReference: "BANK002"
                      payeeProductInstanceReference: "EC98CREDITOR"
                    paymentDetails:
                      amount: 150.75
                      currency: "USD"
                      paymentMechanismType: "CreditTransfer"
                    dateInformation:
                      dateType: "RequestedExecutionDate"
                      date: "2025-10-31"
                    remittanceInformation: "Factura 001-123"
      responses:
        '201':
          description: Orden de pago creada exitosamente
          headers:
            Location:
              description: URI de la orden de pago creada
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiatePaymentOrderResponse'
              examples:
                example1:
                  summary: Respuesta exitosa
                  value:
                    paymentOrderProcedure:
                      paymentOrderProcedureInstanceReference: "550e8400-e29b-41d4-a716-446655440000"
                      paymentOrderProcedureInstanceStatus: "Initiated"
                    _metadata:
                      createdDateTime: "2025-10-30T14:30:00Z"
                    _links:
                      self:
                        href: "/payment-initiation/payment-orders/550e8400-e29b-41d4-a716-446655440000"
                      status:
                        href: "/payment-initiation/payment-orders/550e8400-e29b-41d4-a716-446655440000/status"
        '400':
          description: Request inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /payment-initiation/payment-orders/{paymentOrderId}:
    get:
      tags:
        - Payment Orders
      summary: Recuperar orden de pago completa
      description: |
        Obtiene todos los detalles de una orden de pago existente (Retrieve PaymentOrder).
        Operación nueva no presente en SOAP original.
      operationId: retrievePaymentOrder
      parameters:
        - $ref: '#/components/parameters/PaymentOrderId'
      responses:
        '200':
          description: Orden de pago recuperada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrderResponse'
        '404':
          description: Orden de pago no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /payment-initiation/payment-orders/{paymentOrderId}/status:
    get:
      tags:
        - Payment Orders
      summary: Recuperar estado de orden de pago
      description: |
        Obtiene únicamente el estado actual de una orden de pago (Retrieve PaymentOrder Status).
        Mapea la operación SOAP GetPaymentOrderStatus a REST con estructura BIAN.
      operationId: retrievePaymentOrderStatus
      parameters:
        - $ref: '#/components/parameters/PaymentOrderId'
      responses:
        '200':
          description: Estado recuperado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrderStatusResponse'
        '404':
          description: Orden de pago no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  parameters:
    PaymentOrderId:
      name: paymentOrderId
      in: path
      required: true
      description: Identificador único de la orden de pago (UUID)
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"

  schemas:
    # ==================== REQUEST SCHEMAS ====================
    
    InitiatePaymentOrderRequest:
      type: object
      required:
        - paymentOrderProcedure
      properties:
        paymentOrderProcedure:
          $ref: '#/components/schemas/PaymentOrderProcedureRequest'

    PaymentOrderProcedureRequest:
      type: object
      description: Control Record principal según BIAN 12.0
      required:
        - paymentTransactionInitiatorReference
        - payer
        - payee
        - paymentDetails
        - dateInformation
      properties:
        paymentTransactionInitiatorReference:
          type: string
          description: Referencia del iniciador de la transacción (externalId del SOAP)
          maxLength: 50
          pattern: '^[A-Z0-9-]+$'
          example: "EXT-123"
        payer:
          $ref: '#/components/schemas/Payer'
        payee:
          $ref: '#/components/schemas/Payee'
        paymentDetails:
          $ref: '#/components/schemas/PaymentDetails'
        dateInformation:
          $ref: '#/components/schemas/DateInformation'
        remittanceInformation:
          type: string
          description: Información adicional de remesa
          maxLength: 140
          example: "Factura 001-123"

    Payer:
      type: object
      description: Información del pagador/deudor según BIAN
      required:
        - payerProductInstanceReference
      properties:
        payerReference:
          type: string
          description: Nombre o referencia del pagador
          maxLength: 100
          example: "Juan Pérez"
        payerBankReference:
          type: string
          description: Referencia del banco del pagador
          maxLength: 50
          example: "BANK001"
        payerProductInstanceReference:
          type: string
          description: IBAN o número de cuenta del pagador
          pattern: '^[A-Z]{2}[0-9]{2}[A-Z0-9]{1,30}$'
          example: "EC12DEBTOR"

    Payee:
      type: object
      description: Información del beneficiario/acreedor según BIAN
      required:
        - payeeProductInstanceReference
      properties:
        payeeReference:
          type: string
          description: Nombre o referencia del beneficiario
          maxLength: 100
          example: "María López"
        payeeBankReference:
          type: string
          description: Referencia del banco del beneficiario
          maxLength: 50
          example: "BANK002"
        payeeProductInstanceReference:
          type: string
          description: IBAN o número de cuenta del beneficiario
          pattern: '^[A-Z]{2}[0-9]{2}[A-Z0-9]{1,30}$'
          example: "EC98CREDITOR"

    PaymentDetails:
      type: object
      description: Detalles del pago según BIAN
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          format: double
          description: Monto de la transacción
          minimum: 0.01
          maximum: 999999999.99
          multipleOf: 0.01
          example: 150.75
        currency:
          type: string
          description: Código de moneda ISO 4217
          enum:
            - USD
            - EUR
            - GBP
            - JPY
          example: "USD"
        paymentMechanismType:
          type: string
          description: Tipo de mecanismo de pago
          enum:
            - CreditTransfer
            - DirectDebit
            - CardPayment
            - Check
          example: "CreditTransfer"

    DateInformation:
      type: object
      description: Información de fechas según BIAN
      required:
        - dateType
        - date
      properties:
        dateType:
          type: string
          description: Tipo de fecha
          enum:
            - RequestedExecutionDate
            - ValueDate
            - BookingDate
          example: "RequestedExecutionDate"
        date:
          type: string
          format: date
          description: Valor de la fecha en formato ISO 8601 (YYYY-MM-DD)
          example: "2025-10-31"

    # ==================== RESPONSE SCHEMAS ====================
    
    InitiatePaymentOrderResponse:
      type: object
      required:
        - paymentOrderProcedure
        - _metadata
        - _links
      properties:
        paymentOrderProcedure:
          $ref: '#/components/schemas/PaymentOrderProcedureResponse'
        _metadata:
          $ref: '#/components/schemas/Metadata'
        _links:
          $ref: '#/components/schemas/Links'

    PaymentOrderResponse:
      type: object
      required:
        - paymentOrderProcedure
        - _metadata
        - _links
      properties:
        paymentOrderProcedure:
          $ref: '#/components/schemas/PaymentOrderProcedureFull'
        _metadata:
          $ref: '#/components/schemas/MetadataFull'
        _links:
          $ref: '#/components/schemas/Links'

    PaymentOrderStatusResponse:
      type: object
      required:
        - paymentOrderProcedure
        - _metadata
      properties:
        paymentOrderProcedure:
          type: object
          required:
            - paymentOrderProcedureInstanceReference
            - paymentOrderProcedureInstanceStatus
          properties:
            paymentOrderProcedureInstanceReference:
              type: string
              format: uuid
              description: ID único de la orden de pago
              example: "550e8400-e29b-41d4-a716-446655440000"
            paymentOrderProcedureInstanceStatus:
              $ref: '#/components/schemas/PaymentOrderStatus'
        _metadata:
          type: object
          required:
            - lastUpdateDateTime
          properties:
            lastUpdateDateTime:
              type: string
              format: date-time
              description: Fecha y hora de última actualización
              example: "2025-10-30T16:25:30Z"

    PaymentOrderProcedureResponse:
      type: object
      description: Control Record de respuesta para initiate
      required:
        - paymentOrderProcedureInstanceReference
        - paymentOrderProcedureInstanceStatus
      properties:
        paymentOrderProcedureInstanceReference:
          type: string
          format: uuid
          description: ID único de la orden de pago generado
          example: "550e8400-e29b-41d4-a716-446655440000"
        paymentOrderProcedureInstanceStatus:
          $ref: '#/components/schemas/PaymentOrderStatus'

    PaymentOrderProcedureFull:
      type: object
      description: Control Record completo para retrieve
      required:
        - paymentOrderProcedureInstanceReference
        - paymentTransactionInitiatorReference
        - payer
        - payee
        - paymentDetails
        - dateInformation
        - paymentOrderProcedureInstanceStatus
      properties:
        paymentOrderProcedureInstanceReference:
          type: string
          format: uuid
          description: ID único de la orden de pago
          example: "550e8400-e29b-41d4-a716-446655440000"
        paymentTransactionInitiatorReference:
          type: string
          description: Referencia del iniciador de la transacción
          example: "EXT-123"
        payer:
          $ref: '#/components/schemas/Payer'
        payee:
          $ref: '#/components/schemas/Payee'
        paymentDetails:
          $ref: '#/components/schemas/PaymentDetails'
        dateInformation:
          $ref: '#/components/schemas/DateInformation'
        remittanceInformation:
          type: string
          description: Información de remesa
          example: "Factura 001-123"
        paymentOrderProcedureInstanceStatus:
          $ref: '#/components/schemas/PaymentOrderStatus'

    PaymentOrderStatus:
      type: string
      description: Estado de la orden de pago según BIAN
      enum:
        - Initiated
        - Pending
        - Accepted
        - InProgress
        - Completed
        - Settled
        - Rejected
        - Failed
        - Cancelled
      example: "Initiated"

    Metadata:
      type: object
      description: Metadatos de la respuesta
      required:
        - createdDateTime
      properties:
        createdDateTime:
          type: string
          format: date-time
          description: Fecha y hora de creación
          example: "2025-10-30T14:30:00Z"

    MetadataFull:
      type: object
      description: Metadatos completos de la respuesta
      required:
        - createdDateTime
        - lastUpdateDateTime
      properties:
        createdDateTime:
          type: string
          format: date-time
          description: Fecha y hora de creación
          example: "2025-10-30T14:30:00Z"
        lastUpdateDateTime:
          type: string
          format: date-time
          description: Fecha y hora de última actualización
          example: "2025-10-30T16:25:30Z"

    Links:
      type: object
      description: Enlaces HATEOAS
      required:
        - self
      properties:
        self:
          type: object
          required:
            - href
          properties:
            href:
              type: string
              format: uri
              description: URI del recurso
              example: "/payment-initiation/payment-orders/550e8400-e29b-41d4-a716-446655440000"
        status:
          type: object
          properties:
            href:
              type: string
              format: uri
              description: URI del estado del recurso
              example: "/payment-initiation/payment-orders/550e8400-e29b-41d4-a716-446655440000/status"

    # ==================== ERROR SCHEMAS ====================
    
    ErrorResponse:
      type: object
      description: Respuesta de error según RFC 7807
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI que identifica el tipo de error
          example: "https://api.bank.com/errors/validation-error"
        title:
          type: string
          description: Resumen del error
          example: "Validation Error"
        status:
          type: integer
          description: Código de estado HTTP
          example: 400
        detail:
          type: string
          description: Explicación detallada del error
          example: "El campo payerProductInstanceReference no cumple con el formato IBAN"
        instance:
          type: string
          format: uri
          description: URI que identifica la instancia específica del error
          example: "/payment-initiation/payment-orders"
        timestamp:
          type: string
          format: date-time
          description: Timestamp del error
          example: "2025-10-30T14:30:00Z"